Move interp functions from interp_params to interp plugin
